requirements.txt
requests
feedparser
beautifulsoup4
readability-lxml
Pillow
python-dotenv
